Revert "Player activity added with all functionalities"
