+++ app/src/main/java/com/example/musicplayer_ui/MainActivity.java
@@ -1,6 +1,5 @@
 package com.example.musicplayer_ui;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
@@ -17,8 +16,6 @@ import android.os.Bundle;
 import android.os.Environment;
 import android.provider.MediaStore;
 import android.util.Log;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
@@ -62,8 +59,6 @@ public class MainActivity extends AppCompatActivity {
         super.onDestroy();
         Log.d(TAG,"inside in destroy of main");
     }
-
-
 }
