--- app/src/main/res/drawable/bac_for_pa.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="360dp"
-    android:height="640dp"
-    android:viewportWidth="360"
-    android:viewportHeight="640">
-    <group>
-        <clip-path
-            android:pathData="M0,0h360v640h-360z"/>
-        <path
-            android:pathData="M0,0h360v640h-360z">
-            <aapt:attr name="android:fillColor">
-                <gradient
-                    android:startY="-35"
-                    android:startX="-21.5"
-                    android:endY="800.5"
-                    android:endX="433.5"
-                    android:type="linear">
-                    <item android:offset="0" android:color="#FF3426BB"/>
-                    <item android:offset="1" android:color="#FF6200EE"/>
-                </gradient>
-            </aapt:attr>
-        </path>
-    </group>
-</vector>
