+++ app/src/main/res/layout/activity_main.xml
@@ -4,12 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/background"
+    android:background="#190033"
     tools:context=".MainActivity">
-
     <FrameLayout
-        android:id="@+id/fragment_holder"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent"
+        android:id="@+id/fragment_holder"/>
+
 
 </RelativeLayout>
