--- app/src/main/res/layout/activity_music_player.xml
@@ -1,166 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/bac_for_pa"
-    android:orientation="vertical"
-    android:weightSum="10"
-    tools:context=".MusicPlayerActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="7"
-        android:gravity="center"
-        android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/txtsn"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:ellipsize="marquee"
-            android:marqueeRepeatLimit="marquee_forever"
-            android:padding="10dp"
-            android:singleLine="true"
-            android:layout_marginBottom="40dp"
-            android:text="Song name"
-            android:textAlignment="center"
-            android:textColor="#FFF"
-            android:textSize="18sp"
-            android:typeface="sans">
-
-
-        </TextView>
-
-        <ImageView
-            android:id="@+id/imageview"
-            android:layout_width="180dp"
-            android:layout_height="180dp"
-            android:layout_marginBottom="50dp"
-            android:src="@drawable/music">
-
-        </ImageView>
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="60dp">
-
-            <SeekBar
-                android:id="@+id/seekbar"
-                android:layout_width="250dp"
-                android:layout_height="wrap_content"
-                android:layout_alignParentBottom="true"
-                android:layout_centerInParent="true"
-                android:layout_margin="20dp"
-                android:layout_marginBottom="40dp">
-
-            </SeekBar>
-
-            <TextView
-                android:id="@+id/txtsstart"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentLeft="false"
-                android:layout_centerInParent="true"
-                android:layout_marginLeft="20dp"
-                android:layout_toLeftOf="@+id/seekbar"
-                android:text="0:10"
-                android:textColor="#FFF"
-                android:textSize="12sp">
-
-            </TextView>
-
-            <TextView
-                android:id="@+id/txtsstop"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentRight="false"
-                android:layout_centerInParent="true"
-                android:layout_marginRight="20dp"
-                android:layout_toRightOf="@+id/seekbar"
-                android:text="4:10"
-                android:textColor="#FFF"
-                android:textSize="12sp">
-
-            </TextView>
-
-        </RelativeLayout>
-
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="3">
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginTop="20dp">
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/playbtn"
-                android:layout_width="70dp"
-                android:layout_height="70dp"
-                android:layout_centerHorizontal="true"
-                android:background="@drawable/ic_baseline_pause_circle_24">
-
-            </androidx.appcompat.widget.AppCompatButton>
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btnnext"
-                android:layout_toRightOf="@+id/playbtn"
-                android:layout_marginTop="15dp"
-                android:layout_marginStart="20dp"
-                android:background="@drawable/ic_baseline_skip_next_24"
-                android:layout_width="50dp"
-                android:layout_height="50dp">
-
-            </androidx.appcompat.widget.AppCompatButton>
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btnprev"
-                android:layout_toLeftOf="@+id/playbtn"
-                android:layout_marginTop="15dp"
-                android:layout_marginEnd="20dp"
-                android:background="@drawable/ic_baseline_skip_previous_24"
-                android:layout_width="50dp"
-                android:layout_height="50dp">
-
-            </androidx.appcompat.widget.AppCompatButton>
-
-
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btnff"
-                android:layout_toRightOf="@+id/btnnext"
-                android:layout_marginTop="20dp"
-                android:layout_marginLeft="25dp"
-                android:background="@drawable/ic_baseline_fast_forward_24"
-                android:layout_width="40dp"
-                android:layout_height="40dp">
-
-            </androidx.appcompat.widget.AppCompatButton>
-
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btnfr"
-                android:layout_width="40dp"
-                android:layout_height="40dp"
-                android:layout_marginTop="20dp"
-                android:layout_marginRight="25dp"
-                android:layout_toLeftOf="@+id/btnprev"
-                android:background="@drawable/ic_baseline_fast_rewind_24">
-
-            </androidx.appcompat.widget.AppCompatButton>
-
-
-
-        </RelativeLayout>
-
-    </LinearLayout>
-
-</LinearLayout>
