+++ app/src/main/res/layout/music_item.xml
@@ -1,72 +1,63 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="70dp"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    app:cardCornerRadius="7dp"
-    app:cardBackgroundColor="#513EFD"
-    android:layout_margin="5dp">
+    android:layout_height="60dp"
+    android:layout_marginEnd="10dp"
+    android:layout_marginTop="10dp"
+    android:layout_marginStart="10dp"
+    android:layout_marginBottom="10dp">
 
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:id="@+id/musicItemLayout"
-        android:paddingEnd="11dp"
-        android:padding="5dp"
-        android:gravity="center_vertical">
+        android:background="@drawable/item_bg"
+        android:padding="10dp"
+        android:elevation="10dp">
 
         <ImageView
+            android:layout_width="48dp"
+            android:layout_height="48dp"
             android:id="@+id/music_icon"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
+            android:layout_marginStart="6dp"
+            android:layout_marginTop="6dp"
+            android:background="@drawable/item_bg"
+            android:src="@drawable/ic_music"
             android:layout_alignParentStart="true"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentBottom="true"
-            android:layout_marginLeft="9dp"
-            android:src="@drawable/music" />
+            android:layout_alignParentTop="true"/>
 
         <TextView
             android:id="@+id/music_name"
-            android:layout_width="250dp"
-            android:layout_height="56dp"
+            android:layout_width="wrap_content"
+            android:layout_height="48dp"
             android:layout_alignParentEnd="true"
-            android:layout_marginStart="7dp"
-            android:layout_marginEnd="71dp"
-            android:layout_toEndOf="@+id/music_icon"
+            android:layout_centerInParent="true"
+            android:layout_marginStart="8dp"
+            android:layout_marginEnd="8dp"
+            android:layout_toEndOf="@id/music_icon"
             android:ellipsize="marquee"
             android:fontFamily="sans-serif-medium"
-            android:gravity="center_vertical"
             android:marqueeRepeatLimit="marquee_forever"
             android:padding="8dp"
             android:scrollHorizontally="true"
             android:singleLine="true"
             android:text="Music Name"
-            android:textColor="@color/white"
+            android:textColor="#ffffff"
             android:textSize="15sp" />
 
         <TextView
-            android:id="@+id/music_duration"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_alignParentTop="true"
             android:layout_alignParentEnd="true"
+            android:id="@+id/music_duration"
             android:layout_gravity="end"
-            android:layout_marginTop="36dp"
+            android:layout_marginTop="25dp"
             android:fontFamily="sans-serif"
-            android:gravity="center"
             android:maxLines="1"
             android:text="03.50 min"
-            android:textColor="@color/white"
-            android:textSize="12sp" />
-
-        <ImageView
-            android:id="@+id/settings_img"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentEnd="true"
-            android:layout_marginTop="7dp"
-            app:srcCompat="@drawable/ic_settings" />
+            android:textColor="#ffffff"
+            android:textSize="12sp"
+            android:paddingEnd="10dp"/>
 
 
     </RelativeLayout>
