--- app/src/main/res/menu/fragment_menu.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-<!--    <item-->
-<!--        android:id="@+id/search_action"-->
-<!--        app:actionLayout="@layout/search_widget"-->
-<!--        android:icon="@android:drawable/ic_menu_search"-->
-<!--        android:orderInCategory="1"-->
-<!--        android:title="search"-->
-<!--        app:showAsAction="ifRoom|collapseActionView"/>-->
-    <item
-        android:id="@+id/three_dots"
-        android:icon="@drawable/three_dot_icon"
-        android:orderInCategory="1"
-        android:title="Menu"
-        app:showAsAction="collapseActionView">
-        <menu>
-            <item android:id="@+id/item1"
-                android:title="Sort A-Z" />
-            <item
-                android:id="@+id/item2"
-                android:title="Sort Z-A"/>
-            <item
-                android:id="@+id/item3"
-                android:title="Sort with Duration"/>
-        </menu>
-    </item>
-
-
-
-</menu>
