--- app/src/main/res/values/preloaded_fonts.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <array name="preloaded_fonts" translatable="false">
-        <item>@font/lato</item>
-        <item>@font/lato_bold</item>
-        <item>@font/lato_light</item>
-    </array>
-</resources>
