+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Musgenix</string>
+    <string name="app_name">MusicPlayer_UI</string>
 </resources>
