--- app/src/main/res/xml/file_paths.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<paths xmlns:android="http://schemas.android.com/apk/res/android">
-    <external-path
-        name="external_files"
-        path="."/>
-    <root-path
-        name="external_files"
-        path="/storage/" />
-</paths>
