+++ build.gradle
@@ -5,8 +5,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.2.1'
-
+        classpath "com.android.tools.build:gradle:7.0.3"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
